debug assertation error

+++ Proxi/Proxi/Token.cpp
@@ -7,5 +7,5 @@ Token::Token(TokenType type, std::string lexeme, Literal lit, int line):
 
 std::string Token::toString()
 {
-	return "type: " + type + (std::string)"lexeme: " + lexeme + (std::string)"line: " + std::to_string(line);
+	return "type: " + (int)type + (std::string)"lexeme: " + lexeme + (std::string)"line: " + std::to_string(line);
 }
